Health Check Warning
First time on a vice website checks known HP (Due to API we must wait
for it to be scored first.) If no known HP, nothing happens. When the
player is scored on a vice site we run the same check parsing through
the HP return from the server.

If the users HP is lower than 10, we throw a notification as a warning.